--- src/app/error/page.tsx
@@ -1,38 +1,0 @@
-"use client";
-
-import { useParams, useSearchParams } from "next/navigation";
-
-enum AuthError {
-	Configuration = "Configuration",
-}
-
-const errorMap = {
-	[AuthError.Configuration]: (
-		<p>
-			There was a problem when trying to authenticate. Please contact us if this
-			error persists. Unique error code:{" "}
-			<code className="text-xs bg-slate-100 p-1 rounded-sm">Configuration</code>
-		</p>
-	),
-};
-
-export default function AuthErrorPage() {
-	const search = useSearchParams();
-	const error = search.get("error") as AuthError;
-
-	return (
-		<div className="flex flex-col items-center justify-center w-full h-screen">
-			<a
-				href="/signin"
-				className="block max-w-sm p-6 bg-white border border-gray-200 rounded-lg shadow hover:bg-gray-100 dark:bg-gray-800 dark:border-gray-700 dark:hover:bg-gray-700 text-center"
-			>
-				<h5 className="mb-2 text-xl font-bold tracking-tight text-gray-900 dark:text-white flex flex-row justify-center items-center gap-2">
-					Something went wrong
-				</h5>
-				<div className="font-normal text-gray-700 dark:text-gray-400">
-					{errorMap[error] || "Please contact us if this error persists."}
-				</div>
-			</a>
-		</div>
-	);
-}
