+++ src/app/page.tsx
@@ -1,6 +1,9 @@
 import { Button } from "@/components/ui/button";
 import Image from "next/image";
 import Link from "next/link";
+import nextLogo from "../../public/next.svg";
+
+import { signIn } from "next-auth/react";
 
 export default function Home() {
 	return (
@@ -8,7 +11,7 @@ export default function Home() {
 			<main className="flex flex-col gap-8 row-start-2 items-center sm:items-start">
 				<Image
 					className="dark:invert"
-					src="/next.svg"
+					src={nextLogo}
 					alt="Next.js logo"
 					width={180}
 					height={38}
