+++ src/components/auth/signin-button.tsx
@@ -1,11 +1,16 @@
 "use client";
 
 import { Button } from "@/components/ui/button";
-
 import { signIn } from "next-auth/react";
+import { usePathname } from "next/navigation";
 
 export function SignInButton() {
-	return <Button onClick={() => signIn()}>Sign In</Button>;
+	const pathname = usePathname();
+	return (
+		<Button disabled={pathname === "/signin"} onClick={() => signIn()}>
+			Sign In
+		</Button>
+	);
 }
 
 // export function SignInButton() {
