+++ src/components/auth/signin-form.tsx
@@ -25,6 +25,7 @@ import { signIn } from "next-auth/react";
 import { useRouter } from "next/navigation";
 import { useForm } from "react-hook-form";
 import type { z } from "zod";
+
 export default function SigninForm() {
 	const { toast } = useToast();
 	const router = useRouter();
